sonarcloud

+++ apps/service/models.py
@@ -200,9 +200,8 @@ def auto_delete_file_on_delete(_, instance, **kwargs):
     Deletes file from filesystem
     when corresponding `MediaFile` object is deleted.
     """
-    if instance.video:
-        if os.path.isfile(instance.video.path):
-            os.remove(instance.video.path)
+    if instance.video and os.path.isfile(instance.video.path):
+        os.remove(instance.video.path)
 
 
 class HoroscopeMeme(BaseMeme):
